requirements.txt
# requirements.txt
python-dotenv
langchain
langchain-core             # Contiene Prompts, Output Parsers e Core Runnables
langchain-google-genai
langchain-mongodb
pypdf
langchain-community
langchain-text-splitters